scheduler-service/requirements.txt
kafka-python==2.0.2
redis==5.0.1
prometheus-client==0.17.1
python-dotenv==1.0.0
schedule==1.2.0 